requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.7
django-celery-beat==2.3.0
django-cron==0.6.0
django-filter==22.1
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-timezone-field==5.0
djangorestframework==3.14.0
idna==3.4
importlib-metadata==4.12.0
kombu==5.2.4
Markdown==3.4.1
packaging==21.3
prompt-toolkit==3.0.31
pycodestyle==2.9.1
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.1
